requirements.txt
django
djangorestframework
djangorestframework_simplejwt
pillow
requests
django-cors-headers
python-decouple